Port extant, relevant tree queue tests to new Tree, and finish ripping previous tree queue iteration
