+++ pkg/scheduler/queue/queue_test.go
@@ -671,7 +671,7 @@ func TestRequestQueue_tryDispatchRequestToQuerier_ShouldReEnqueueAfterFailedSend
 
 	require.Nil(t, queueBroker.queueTree.rootNode.getNode(QueuePath{"tenant-1"}))
 	require.NoError(t, queueBroker.enqueueRequestBack(&tr, tenantMaxQueriers))
-	require.False(t, queueBroker.queueTree.rootNode.getNode(QueuePath{"tenant-1"}).IsEmpty())
+	require.False(t, queueBroker.queueTree.GetNode(QueuePath{"tenant-1"}).IsEmpty())
 
 	ctx, cancel := context.WithCancel(context.Background())
 	call := &waitingQuerierConn{
@@ -686,5 +686,5 @@ func TestRequestQueue_tryDispatchRequestToQuerier_ShouldReEnqueueAfterFailedSend
 	// indicating not to re-submit a request for waitingQuerierConn for the querier
 	require.True(t, queue.trySendNextRequestForQuerier(call))
 	// assert request was re-enqueued for tenant after failed send
-	require.False(t, queueBroker.queueTree.rootNode.getNode(QueuePath{"tenant-1"}).IsEmpty())
+	require.False(t, queueBroker.queueTree.GetNode(QueuePath{"tenant-1"}).IsEmpty())
 }
